match: revert 634245a and 8dcf69c and defer tag command to later
slurping actually requires knowledge of which command we are running,
because it behaves differently (when creating the tag message fields),
so we cannot run the slurp prior to command selection as these changes
introduced would have done.  instead we revert them and make a new
deferred command which is resolved later during command execution
(ie during call()), once we have done the slurp and can figure out
successfully if prefix and branch differ (ie rebase-fork case).