/ tagvalids: fix dash in pattern being confused as a range
no idea how this worked for years and suddenly stop working.  makes
no sense.  bash must have changed... but the weird part is, it works on
another system -- just tested -- with same bash version.  so maybe bash
is using some library for REs (like, from libc) that got upgraded and
behaves differently now.  don't know but this is really bizarre.

however, need this to work, it makes sense why the pattern is invalid,
and don't have time to figure out why it changed at the moment...