+ main: implement a dirty flag to allow tagging while changes pending
this comes up frequently, and the usual answer is to stash, reltag, then
unstash, but this is annoying, and also causes the editor to think files
have changed.  if we know already it's not an issue we can give --dirty

along the way we implemented option processing, so there can be more
options in the future.  since we already supported --help we sort of had
to do this because we can't just test $1, we'd have to handle 4
different permutations.  arg processing is just easier, and we already
had a template for it in devskel