- init: allow to specify start vers for init by args like for exact
only for init do we allow previous tags not to exist.  if we have done
some history editing, possibly making a whole new history from the root
commit, then "git describe" will find nothing and 'exact' will fail,
because it bombs when no previous commit (only allowed for 'init').  in
this case we need to specify as though for 'exact' but using the 'init'
logic of not requiring a previous tag to exist (because basically we're
starting a new tag lineage, albeit numerically contiguous with previous
tags)